+++ Services/OrderService.cs
@@ -35,8 +35,8 @@ namespace Nwazet.Commerce.Services {
         public OrderPart CreateOrder(
             ICharge charge,
             IEnumerable<CheckoutItem> items,
-            double subTotal,
-            double total,
+            decimal subTotal,
+            decimal total,
             TaxAmount taxes,
             ShippingOption shippingOption,
             Address shippingAddress,
@@ -48,7 +48,7 @@ namespace Nwazet.Commerce.Services {
             string trackingUrl = null,
             bool isTestOrder = false,
             int userId = -1,
-            double amountPaid = 0,
+            decimal amountPaid = 0,
             string purchaseOrder = "") {
 
             var order = _contentManager.Create("Order", VersionOptions.DraftRequired).As<OrderPart>();
